Calibrate stage-shift based survival to PCBaSe
The stage-shift model is now set as the default survival model. The
stage-shift model has been calibrated to the survival in PCBaSe. This was
done using prostate cancer specific survival at 10 and 15 years since
diagnosis for all treatments.

The choice to calibrate to all treatments was made to avoid bias by
confounding-by-treatment (sicker patients might be over-represented in
conservatively/non-curatively treated). The HR from the two time points
were averaged on the log-scale.